refactor: extract service duration helpers
